+++ tests/measurements/test_mid_measure.py
@@ -1,0 +1,31 @@
+# Copyright 2018-2020 Xanadu Quantum Technologies Inc.
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+#     http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""Unit tests for the mid_measure module"""
+
+import pytest
+
+import pennylane as qml
+
+
+class TestMeasure:
+    """Tests for the measure function"""
+
+    def test_many_wires_error(self):
+        """Test that an error is raised if multiple wires are passed to
+        measure."""
+        with pytest.raises(
+            qml.QuantumFunctionError,
+            match="Only a single qubit can be measured in the middle of the circuit",
+        ):
+            qml.measure(wires=[0, 1])
